third_party/WebKit/LayoutTests/fullscreen/api/document-exit-fullscreen-twice.html

<!DOCTYPE html>
<title>Document#exitFullscreen() called twice</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
// Adapted from https://github.com/w3c/web-platform-tests/pull/4250
// TODO(foolip): Remove this test when the above is imported and passing.
async_test(t => {
  const div = document.querySelector("div");

  document.onfullscreenchange = t.step_func(() => {
    // We are now in fullscreen.
    assert_equals(document.fullscreenElement, div);

    document.onfullscreenchange = t.step_func(() => {
      assert_equals(document.fullscreenElement, null);
      // Done, but ensure that there's only one fullscreenchange event.
      document.onfullscreenchange = t.unreached_func("second fullscreenchange event");
      setTimeout(t.step_func_done(), 0);
    });

    // Exit fullscreen twice.
    document.exitFullscreen();
    assert_equals(document.fullscreenElement, null, "fullscreenElement after first exitFullscreen()");
    document.exitFullscreen();
    assert_equals(document.fullscreenElement, null, "fullscreenElement after second exitFullscreen()");
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_request(div);
});
</script>
